dom/base/test/jsmodules/test_multiTopLevelLargeImports.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Test a importing large modules which may be compiled off main thread</title>
<script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
<script>
  var results = [];

  SimpleTest.waitForExplicitFinish();

  function arrayEquals(a, b) {
    if (a.length != b.length) {
      return false;
    }
    for (let i = 0; i < a.length; i++) {
      if (a[i] != b[i]) {
        return false;
      }
      return true;
    }
  }

  function testLoaded() {
    ok(arrayEquals(results, [1, 2, 3]), 'Check modules imported');
    SimpleTest.finish();
  }
</script>
<script type="module" src="module_large1.js"></script>
<script type="module" src="module_simple2.js"></script>
<script type="module" src="module_large3.js"></script>
<body onload='testLoaded()'></body>
